+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and added `String` field `code`
 - **Breaking**: Renamed `Organization.VirtualWitchPhenomenon` to `Organization.KAMITSUBAKI`
 - **Breaking**: `EnumUtil` is no longer exported.
 - **Breaking**: Changed `Language` to a class to support the `BCP47` standard.
+- **Breaking**: Changed `ChannelMin` to use `ChannelStats` instead of multiple related fields on `Channel`.
 - Changed `Video` constructor parameter `duration` to optional
 have been replaced with `VideoLiveInfo`
 - Added `Serializable`